backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
alembic==1.13.0
google-generativeai==0.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosmtplib==3.0.1
slowapi==0.1.8
pytest==7.4.3
httpx==0.25.2
pandas==2.1.3
Pillow==10.1.0
sentry-sdk==1.38.0
